requirements.txt
streamlit
tensorflow
'''numpy>=1.21.0
pandas>=1.3.0
Pillow>=9.0.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
h5py>=3.1.0
protobuf<4.0.0 
'''